Separate native helper into sub-directory
